Update migrated_content.md
For the introduction, Review of Code Review Guide 2.0, and licensing there were duplicate paragraphs and titles. I removed all of the duplicates because they aren't supposed to be there. Also I feel the title "Review of Code Review Guide 2.0" should be changed to "Code Review Guide 2.0" or "Review of Code Guide 2.0". I feel "Code Review Guide 2.0" sounds the best so that is what I propose it gets changed to. I made sure to change the id to "code_review_guide_2.0" as well.